requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==1.0.8
Werkzeug==0.10.4
bcrypt==2.0.0
cffi==1.2.1
fudge==1.0.3
itsdangerous==0.24
pycparser==2.14
requests==2.8.0
simplejson==3.8.0
six==1.10.0
soundcloud==0.4.1
wsgiref==0.1.2
youtube-dl==2015.10.09
